--- packages/@tests-integration/src/__test__/integration/marine-js.spec.d.ts
@@ -1,2 +1,0 @@
-export {};
-//# sourceMappingURL=marine-js.spec.d.ts.map
