+++ packages/@tests-integration/src/__test__/unit/KeyPair.spec.ts
@@ -1,6 +1,5 @@
 import * as bs58 from 'bs58';
-import * as base64 from 'base64-js';
-import { KeyPair } from '../../internal/KeyPair';
+import { KeyPair } from '@fluencelabs/fluence';
 
 describe('KeyPair tests', () => {
     it('generate keypair from seed', async function () {
