+++ packages/@tests-integration/src/_aqua/services.ts
@@ -7,132 +7,131 @@
  *
  */
 import { Fluence, FluencePeer } from '@fluencelabs/fluence';
-import {
-    CallParams,
-    callFunction,
-    registerService,
-} from '@fluencelabs/fluence/dist/internal/compilerSupport/v3';
-
+import { CallParams, callFunction, registerService } from '@fluencelabs/fluence/dist/internal/compilerSupport/v3';
 
 // Services
 
 export interface SigDef {
     get_pub_key: (callParams: CallParams<null>) => string | Promise<string>;
-    sign: (data: number[], callParams: CallParams<'data'>) => { error: string | null; signature: number[] | null; success: boolean; } | Promise<{ error: string | null; signature: number[] | null; success: boolean; }>;
-    verify: (signature: number[], data: number[], callParams: CallParams<'signature' | 'data'>) => boolean | Promise<boolean>;
+    sign: (
+        data: number[],
+        callParams: CallParams<'data'>,
+    ) =>
+        | { error: string | null; signature: number[] | null; success: boolean }
+        | Promise<{ error: string | null; signature: number[] | null; success: boolean }>;
+    verify: (
+        signature: number[],
+        data: number[],
+        callParams: CallParams<'signature' | 'data'>,
+    ) => boolean | Promise<boolean>;
 }
 export function registerSig(service: SigDef): void;
 export function registerSig(serviceId: string, service: SigDef): void;
 export function registerSig(peer: FluencePeer, service: SigDef): void;
 export function registerSig(peer: FluencePeer, serviceId: string, service: SigDef): void;
-       
 
 export function registerSig(...args: any) {
-    registerService(
-        args,
-        {
-    "defaultServiceId" : "sig",
-    "functions" : {
-        "tag" : "labeledProduct",
-        "fields" : {
-            "get_pub_key" : {
-                "tag" : "arrow",
-                "domain" : {
-                    "tag" : "nil"
+    registerService(args, {
+        defaultServiceId: 'sig',
+        functions: {
+            tag: 'labeledProduct',
+            fields: {
+                get_pub_key: {
+                    tag: 'arrow',
+                    domain: {
+                        tag: 'nil',
+                    },
+                    codomain: {
+                        tag: 'unlabeledProduct',
+                        items: [
+                            {
+                                tag: 'scalar',
+                                name: 'string',
+                            },
+                        ],
+                    },
                 },
-                "codomain" : {
-                    "tag" : "unlabeledProduct",
-                    "items" : [
-                        {
-                            "tag" : "scalar",
-                            "name" : "string"
-                        }
-                    ]
-                }
-            },
-            "sign" : {
-                "tag" : "arrow",
-                "domain" : {
-                    "tag" : "labeledProduct",
-                    "fields" : {
-                        "data" : {
-                            "tag" : "array",
-                            "type" : {
-                                "tag" : "scalar",
-                                "name" : "u8"
-                            }
-                        }
-                    }
-                },
-                "codomain" : {
-                    "tag" : "unlabeledProduct",
-                    "items" : [
-                        {
-                            "tag" : "struct",
-                            "name" : "SignResult",
-                            "fields" : {
-                                "error" : {
-                                    "tag" : "option",
-                                    "type" : {
-                                        "tag" : "scalar",
-                                        "name" : "string"
-                                    }
+                sign: {
+                    tag: 'arrow',
+                    domain: {
+                        tag: 'labeledProduct',
+                        fields: {
+                            data: {
+                                tag: 'array',
+                                type: {
+                                    tag: 'scalar',
+                                    name: 'u8',
                                 },
-                                "signature" : {
-                                    "tag" : "option",
-                                    "type" : {
-                                        "tag" : "array",
-                                        "type" : {
-                                            "tag" : "scalar",
-                                            "name" : "u8"
-                                        }
-                                    }
-                                },
-                                "success" : {
-                                    "tag" : "scalar",
-                                    "name" : "bool"
-                                }
-                            }
-                        }
-                    ]
-                }
-            },
-            "verify" : {
-                "tag" : "arrow",
-                "domain" : {
-                    "tag" : "labeledProduct",
-                    "fields" : {
-                        "signature" : {
-                            "tag" : "array",
-                            "type" : {
-                                "tag" : "scalar",
-                                "name" : "u8"
-                            }
+                            },
                         },
-                        "data" : {
-                            "tag" : "array",
-                            "type" : {
-                                "tag" : "scalar",
-                                "name" : "u8"
-                            }
-                        }
-                    }
+                    },
+                    codomain: {
+                        tag: 'unlabeledProduct',
+                        items: [
+                            {
+                                tag: 'struct',
+                                name: 'SignResult',
+                                fields: {
+                                    error: {
+                                        tag: 'option',
+                                        type: {
+                                            tag: 'scalar',
+                                            name: 'string',
+                                        },
+                                    },
+                                    signature: {
+                                        tag: 'option',
+                                        type: {
+                                            tag: 'array',
+                                            type: {
+                                                tag: 'scalar',
+                                                name: 'u8',
+                                            },
+                                        },
+                                    },
+                                    success: {
+                                        tag: 'scalar',
+                                        name: 'bool',
+                                    },
+                                },
+                            },
+                        ],
+                    },
                 },
-                "codomain" : {
-                    "tag" : "unlabeledProduct",
-                    "items" : [
-                        {
-                            "tag" : "scalar",
-                            "name" : "bool"
-                        }
-                    ]
-                }
-            }
-        }
-    }
+                verify: {
+                    tag: 'arrow',
+                    domain: {
+                        tag: 'labeledProduct',
+                        fields: {
+                            signature: {
+                                tag: 'array',
+                                type: {
+                                    tag: 'scalar',
+                                    name: 'u8',
+                                },
+                            },
+                            data: {
+                                tag: 'array',
+                                type: {
+                                    tag: 'scalar',
+                                    name: 'u8',
+                                },
+                            },
+                        },
+                    },
+                    codomain: {
+                        tag: 'unlabeledProduct',
+                        items: [
+                            {
+                                tag: 'scalar',
+                                name: 'bool',
+                            },
+                        ],
+                    },
+                },
+            },
+        },
+    });
 }
-    );
-}
-      
-// Functions
 
+// Functions
