+++ packages/fluence-js/src/__test__/compiler/v2.spec.ts.ignore
@@ -1,7 +1,7 @@
 import each from 'jest-each';
 
-import { Fluence, FluencePeer } from '../../..';
-import { forTests } from '../../../internal/compilerSupport/v2';
+import { Fluence, FluencePeer } from '../..';
+import { forTests } from '../../internal/compilerSupport/v2';
 
 const peer = new FluencePeer();
 const cfg = { ttl: 1000 };
