+++ packages/fluence-js/src/services.ts
@@ -1,2 +1,3 @@
 export * from './internal/builtins/Sig';
 export { registerSig } from './internal/_aqua/services';
+export type { SigDef } from './internal/_aqua/services';
